ghost_blogger_agent/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghost-blogger-agent"
version = "0.1.0"
description = "GitHub-native web-reading agent that writes reflective Jekyll blog posts using a tiny local model."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "beautifulsoup4>=4.12",
  "feedparser>=6.0",
  "httpx>=0.27",
  "lxml>=5.2",
  "python-dateutil>=2.9",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]
llm = [
  "torch>=2.2; platform_system != 'Darwin' or platform_machine != 'arm64'",
]

[project.scripts]
ghost-blogger = "ghost_blogger.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
